requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
python-dotenv
jinja2
requests
